Python/simple-root-bot/requirements.txt
botbuilder-core==v4.7.1
aiohttp
